Added FrontDisplayEmulator. Moved some inline classes/enums to separate files. SD Card mount: wait while RemovableMedia.Inserted fired. Refactor reset functionality: fire reset only after playing stopped(history was saved). Refactor layout of OnBoardMonitorEmulator windows application. Added phone indicators handler for emulator app. Added separate 'readBufferSize' for imBMW and emulator applications. Replace some int variables to byte or ushort. Added displaying of track title on OBC display of instrument cluster. Header message was send only if it contains text. Fixed infinite loop on MenuItem.Text property. Added periodical refresh of bordmonitor header with one symbol shifting. Add OnIsPlayingChanging event. Removed CDChanger answering of disk and track number, when track changing. Added IsNullOrEmpty extension. Refactor 'GetDateTime' method for avoiding null returns, increased wait time to 2sec. Added some new 'messageDescriptions'

+++ Sources/NET-MF/Devices/imBMW.UnitTests/TestProgram.cs
@@ -15,7 +15,7 @@ namespace imBMW.UnitTests
     {
         public void ShouldTrimText_WithLength_LessThanTrimLength()
         {
-            int startIndex;
+            byte startIndex;
             string testString = "Dua Lipa";
             int trimLength = 10;
             string result = "";
@@ -30,7 +30,7 @@ namespace imBMW.UnitTests
 
         public void ShouldTrimText_WithLength_EqualToTrimLength()
         {
-            int startIndex;
+            byte startIndex;
             string testString = "Dua Lipa";
             int trimLength = testString.Length;
             string result = "";
@@ -45,7 +45,7 @@ namespace imBMW.UnitTests
 
         public void ShouldTrimText_WithLength_BiggerThanTrimLength()
         {
-            int startIndex;
+            byte startIndex;
             string testString = "The Last Of The Real Ones";
             int trimLength = 10;
             string result = "";
