+++ Sources/NET-MF/imBMW.Features/Menu/MenuHelpers.cs
@@ -6,7 +6,7 @@ namespace imBMW.Features.Menu
 {
     public static class MenuHelpers
     {
-        public static void AddBackButton(this MenuScreen screen, int index = -1)
+        public static void AddBackButton(this MenuScreen screen, SByte index = -1)
         {
             screen.AddItem(new MenuItem(i => "« " + Localization.Current.Back, MenuItemType.Button, MenuItemAction.GoBackScreen), index);
         }
